Specialize slider components and refactor shared logic
- Created useBaseSelectionSlider hook for shared numerical input synchronization
- Restored useOptionSlider to handle index-based option mapping
- Refactored OptionSliderControl to use domain-specific hooks and props
- Implemented SimulationSliderControl with a value-based API (min, max, value)
- Unified visual layout via BaseSelectionSliderLayout component
- Cleaned up SimulationDialog by removing index mapping arrays
- Ensured architectural separation while maintaining UI consistency
- Fixed Biome linting and import sorting issues
